refactor(OutputMessage): Class 에서 Enum으로 변환 및 변화에 의한 수정
