+++ src/main/java/christmas/constants/OutputMessage.java
@@ -1,9 +1,8 @@
-package christmas.view;
+package christmas.constants;
 
 public enum OutputMessage {
 
-    REQUEST_DATE_MESSAGE("12월 중 식당 예상 방문 날짜는 언제인가요? (숫자만 입력해 주세요!)"),
-    ;
+    REQUEST_DATE_MESSAGE("12월 중 식당 예상 방문 날짜는 언제인가요? (숫자만 입력해 주세요!)");
 
 
     private final String message;
